heavy re-org to make player testable
